Code refactoring for server package
